requirements.txt
defusedxml==0.7.1
fpdf2==2.5.5
Pillow==9.2.0
python-barcode==0.14.0
qrcode==7.3.1
